requirements.txt
# general modules
django 
Pillow
djangorestframework


# third party modules
markdown     
django-filter
coreapi
drf-yasg[validation]
djoser


# deployment modules


# email third party modules
django-mail-templated

# refactoring and tests
black
flake8